first Commit
This code has an Nrf24 communication, it is able to put up two led with a potentiometer at various range. it also give reply back to the sensing node which led is on or off. The challenge I am facing is that when the system works for a time, there seems to be conflict in transmission and the led bulbs start coming on and off at will.